+++ engine/interface/pos.cpp
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 // interface.cpp
 // interface rendering implementation
-// $Id: pos.cpp,v 1.2 2003/11/18 18:39:42 tstivers Exp $
+// $Id: pos.cpp,v 1.3 2003/12/03 07:21:39 tstivers Exp $
 //
 
 #include "precompiled.h"
@@ -43,13 +43,13 @@ void Pos::render()
 	}
 
 	sprintf(buf, "pos x: %.2f y: %.2f z: %.2f", game::player.pos.x, game::player.pos.y, game::player.pos.z);
-	d3dfont->DrawText(xpos, ypos, D3DCOLOR_XRGB(255, 255, 255), buf);
+	d3dfont->DrawText((float)xpos, (float)ypos, D3DCOLOR_XRGB(255, 255, 255), buf);
 	sprintf(buf, "rot x: %.2f y: %.2f z: %.2f", game::player.rot.x, game::player.rot.y, game::player.rot.z);
-	d3dfont->DrawText(xpos, ypos + 10, D3DCOLOR_XRGB(255, 255, 255), buf);
+	d3dfont->DrawText((float)xpos, (float)ypos + 10, D3DCOLOR_XRGB(255, 255, 255), buf);
 	sprintf(buf, "vel x: %.2f y: %.2f z: %.2f", game::player.vel.x, game::player.vel.y, game::player.vel.z);
-	d3dfont->DrawText(xpos, ypos + 20, D3DCOLOR_XRGB(255, 255, 255), buf);
+	d3dfont->DrawText((float)xpos, (float)ypos + 20, D3DCOLOR_XRGB(255, 255, 255), buf);
 	if(q3bsp::bsp) {
 		sprintf(buf, "leafs: %i faces: %i polys: %i texswaps: %i", q3bsp::bsp->frame_leafs, q3bsp::bsp->frame_faces, q3bsp::bsp->frame_polys, q3bsp::bsp->frame_textureswaps);
-		d3dfont->DrawText(xpos, ypos + 30, D3DCOLOR_XRGB(255, 255, 255), buf);
+		d3dfont->DrawText((float)xpos, (float)ypos + 30, D3DCOLOR_XRGB(255, 255, 255), buf);
 	}
 }
